Move event definitions from app to proxy
This avoids a circular reference during the tests.
